+++ Changelog.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Añadido
+- Subreddits LMDShow, MemesESP y orslokx en memes
+- Los memes pueden ser ahora videos
+
 ### Eliminado
 - Comando count_memes
 - Cog ReactionRoles
+- Subreddit furry_irl en memes
 
 ## v4.3.4 - 2022-3-28
 ### Añadido